requirements.txt
Django==3.2
djangorestframework==3.12.4
channels==3.0.3
channels-redis==3.2.0
shortuuid==1.0.1
redis==3.5.3
django-cors-headers==3.7.0
